Require an opt in packet for networked workspace editors
