+++ src/main/java/xyz/nucleoid/creator_tools/CreatorTools.java
@@ -4,12 +4,15 @@ import com.google.common.reflect.Reflection;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerTickEvents;
+import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import xyz.nucleoid.creator_tools.command.MapManageCommand;
 import xyz.nucleoid.creator_tools.command.MapMetadataCommand;
 import xyz.nucleoid.creator_tools.item.CreatorToolsItems;
 import xyz.nucleoid.creator_tools.workspace.MapWorkspaceManager;
+import xyz.nucleoid.creator_tools.workspace.WorkspaceTraveler;
+import xyz.nucleoid.creator_tools.workspace.editor.WorkspaceNetworking;
 
 public final class CreatorTools implements ModInitializer {
     public static final String ID = "nucleoid_creator_tools";
@@ -27,6 +30,11 @@ public final class CreatorTools implements ModInitializer {
 
         ServerTickEvents.START_SERVER_TICK.register(server -> {
             MapWorkspaceManager.get(server).tick();
+        });
+
+        ServerPlayNetworking.registerGlobalReceiver(WorkspaceNetworking.OPT_IN_ID, (server, player, handler, buf, responseSender) -> {
+            int protocolVersion = buf.readVarInt();
+            WorkspaceTraveler.setCreatorToolsProtocolVersion(player, protocolVersion);
         });
     }
 }
