+++ src/main/java/xyz/nucleoid/creator_tools/workspace/WorkspaceTraveler.java
@@ -3,6 +3,7 @@ package xyz.nucleoid.creator_tools.workspace;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.registry.RegistryKey;
 import net.minecraft.world.World;
+import xyz.nucleoid.creator_tools.workspace.editor.WorkspaceNetworking;
 import org.jetbrains.annotations.Nullable;
 
 public interface WorkspaceTraveler {
@@ -22,9 +23,26 @@ public interface WorkspaceTraveler {
         return null;
     }
 
+    static int getCreatorToolsProtocolVersion(ServerPlayerEntity player) {
+        if (player instanceof WorkspaceTraveler traveler) {
+            return traveler.getCreatorToolsProtocolVersion();
+        }
+        return WorkspaceNetworking.NO_PROTOCOL_VERSION;
+    }
+
+    static void setCreatorToolsProtocolVersion(ServerPlayerEntity player, int protocolVersion) {
+        if (player instanceof WorkspaceTraveler traveler) {
+            traveler.setCreatorToolsProtocolVersion(protocolVersion);
+        }
+    }
+
     @Nullable
     ReturnPosition getReturnFor(RegistryKey<World> dimension);
 
     @Nullable
     ReturnPosition getLeaveReturn();
+
+    int getCreatorToolsProtocolVersion();
+
+    void setCreatorToolsProtocolVersion(int protocolVersion);
 }
