+++ src/main/java/xyz/nucleoid/creator_tools/workspace/editor/WorkspaceNetworking.java
@@ -7,6 +7,8 @@ import xyz.nucleoid.creator_tools.workspace.WorkspaceRegion;
 import xyz.nucleoid.map_templates.BlockBounds;
 
 public final class WorkspaceNetworking {
+    public static final int NO_PROTOCOL_VERSION = -1;
+
     // Client <-> Server
     public static final Identifier WORKSPACE_LEAVE_ID = new Identifier(CreatorTools.ID, "workspace/leave");
     public static final Identifier WORKSPACE_BOUNDS_ID = new Identifier(CreatorTools.ID, "workspace/bounds");
@@ -19,6 +21,7 @@ public final class WorkspaceNetworking {
     public static final Identifier WORKSPACE_REGIONS_ID = new Identifier(CreatorTools.ID, "workspace/regions");
 
     // Client --> Server
+    public static final Identifier OPT_IN_ID = new Identifier(CreatorTools.ID, "opt_in");
     public static final Identifier WORKSPACE_NEW_ID = new Identifier(CreatorTools.ID, "workspace/new");
     public static final Identifier WORKSPACE_REGION_ADD_ID = new Identifier(CreatorTools.ID, "workspace/region/add");
